Remove webhooks and refactor websockets
