+++ src/routes/issuer.js
@@ -8,7 +8,7 @@ import { countIssuers, readIssuerById } from "../db/operations/read.js";
 import { convertUUIDToBytes16 } from "../utils/convertUUID.js";
 import validateInputAgainstOCF from "../utils/validateInputAgainstSchema.js";
 import { checkPortal } from "../fairmint/checkPortal.js";
-import { addAddressesToWatch } from "../websocket.mjs";
+import { addAddressesToWatch } from "../utils/websocket.js";
 const issuer = Router();
 
 issuer.get("/", async (req, res) => {
